pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "newscale"
description = "Control logic for the New Scale micromanipulators."
license = {text = "MIT"}
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3"
]
dynamic = ["version", "readme"]

dependencies = [
    'pyserial >= 3.5'
]

[project.optional-dependencies]
dev = [
    'black',
    'coverage',
    'flake8',
    'interrogate',
    'isort',
    'Sphinx'
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "newscale.__version__"}
readme = {file = ["README.md"]}

[tool.black]
line-length = 79
target_version = ['py37']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
  )/
  | .gitignore
)
'''

[tool.coverage.run]
omit = ["*__init__*"]
source = ["newscale", "tests"]

[tool.coverage.report]
exclude_lines = ["if __name__ == .__main__.:"]
fail_under = 100

[tool.isort]
line_length = 79

[tool.interrogate]
exclude = ["setup.py", "docs", "build"]
fail-under = 100
